test: isolate global panic-capture coverage in its own process
The `init_global_installs_panic_capture_by_default` unit test initialized
the real process-wide `GLOBAL_CLIENT` (a set-once `OnceLock` that can't be
reset), which made it order-dependent: any future test that touched the
global, or a reordering, could break the suite. It can't be fixed with
dependency injection either — the panic hook needs a `'static` client, so
only the global `OnceLock` qualifies and there's nothing to mock.

Move that end-to-end check to its own integration-test binary
(`tests/test_panic_global.rs`), where the `OnceLock` is naturally isolated,
and drop the in-process test plus its now-unused `init_global_test` helper.
The gate (`should_capture_global_panics`) and the hook mechanism
(`install_panic_hook`) stay covered by the in-process unit tests; the
integration test only exercises the `init_global` -> hook wiring, and runs
across all error-tracking feature configs (V0/V1 x async/blocking).
